% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iwgvemm.R
\name{iwgvemm}
\alias{iwgvemm}
\title{IW-GVEMM Algorithm for DIF Detection in 2PL Models}
\usage{
iwgvemm(
  Y,
  D,
  X,
  Lambda0 = seq(0.2, 0.7, by = 0.1),
  S = 10,
  M = 10,
  lr = 0.1,
  iter = 1000,
  eps = 0.001,
  c = 0.7
)
}
\arguments{
\item{Y}{An N by J binary matrix of item responses}

\item{D}{A J by G binary matrix of loading indicators}

\item{X}{An N dimensional vector of group indicators (integers from 1 to G)}

\item{Lambda0}{A vector of \code{lambda0} values for L1 penalty (\code{lambda} is \code{sqrt(N) * lambda0})}

\item{S}{Sample size for approximating the expected lower bound}

\item{M}{Sample size for approximating a tighter lower bound}

\item{lr}{Learning rate for Adam}

\item{iter}{Maximum number of iterations}

\item{eps}{Termination criterion on numerical accuracy}

\item{c}{Constant for computing GIC}
}
\value{
A list whose length is equal to \code{Lambda0}
\item{lambda0}{Corresponding element in \code{Lambda0}}
\item{lambda}{\code{sqrt(N) * lambda0}}
\item{SIGMA}{Person-level posterior covariance matrices}
\item{MU}{Person-level posterior mean vectors}
\item{Sigma}{Group-level posterior covariance matrices}
\item{Mu}{Group-level posterior mean vectors}
\item{a}{Slopes for group 1}
\item{b}{Intercepts for group 1}
\item{gamma}{DIF parameters for the slopes}
\item{beta}{DIF parameters for the intercepts}
\item{ll}{Log-likelihood}
\item{l0}{Number of nonzero parameters in \code{gamma} and \code{beta}}
\item{AIC}{Akaike Information Criterion}
\item{BIC}{Bayesian Information Criterion}
\item{GIC}{Generalized Information Criterion}
}
\description{
IW-GVEMM Algorithm for DIF Detection in 2PL Models
}
\examples{
with(gvemm_simdata) iwgvemm(Y, D, X)
}
